// src/app/core/model/user-devotee.model.ts
export class UserDevotee {
  profileImg: string;
  othersImg1: string;
  othersImg2: string;
  othersImg3: string;
  first_name: string;
  last_name: string;
  email: string;
  phone: number;
  phoneVisibility: string;
  profession: string;
  genre: string;
  orientation: string;
  aboutYou: string = '';
  checkImpairedFertility: boolean = false;
  checkBirthDisability: boolean = false;
  checkHaveTIIc: boolean = false;
  checkOvercomingExample: boolean = false;
  surgeries: string[];
  hospitals: string[];
  medicines: string[];
}